ClientAPI/requirements.txt
aiofiles==0.5.0
click==7.1.2
dataclasses==0.7
dnspython==2.0.0
fastapi==0.61.1
h11==0.9.0
httptools==0.1.1
Pillow==7.2.0
pydantic==1.6.1
pymongo==3.11.0
starlette==0.13.6
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
